+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Added `geometry.find_coords_on_rect()`, which returns coordinates along the perimeter of a rectangle given a center `Coord`, width, and height. Results are cached for performance.
 * `Coord` objects can be unpacked into `(column, row)` tuples for multiple assignment.
 * `motion.activate_path()` and `animation.activate_scene` accept `path_id`/`scene_id` strings OR `Path`/`Scene` instances. The corresponding `Path`/`Scene` must exist.
+* Added `--terminal-background-color` to the `TerminalConfig` parser. This will enable terminal themes with background other than black to better display effects with fade in/out components.
 
 ---